+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
-## [v2.2.17.dev0]
+## [v2.3.1.dev0]
+
+### Added
+- 
+
+### Changed
+- 
+
+### Fixes
+- [Cli] Fixed --config param in lithops cli
+- [Standalone] Fixed internal executions
+
+
+## [v2.3.0]
 
 ### Added
 - [Core] Added multiprocessing support in workers